requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
pystray>=0.19.4
Pillow>=10.0.0
matplotlib>=3.7.0
pyinstaller>=5.13.0 